pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "geo-mlops"
version = "0.1.0"
description = "Contract-driven geospatial MLOps pipeline for remote sensing ML workflows."
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Mouhamad Ali Elamine" }
]
dependencies = [
    "mlflow>=3.11,<4",
    "numpy>=2.0,<3",
    "pandas>=2.0,<4",
    "psutil>=7.0,<8",
    "pydantic>=2.0,<3",
    "pyproj>=3.7,<4",
    "pyyaml>=6.0,<7",
    "rasterio>=1.4,<2",
    "tqdm>=4.0,<5",
]

[project.optional-dependencies]
torch = [
    "torch>=2.0",
    "torchvision>=0.20",
]

models = [
    "transformers>=4.40,<6",
]

ml = [
    "torch>=2.0",
    "torchvision>=0.20",
    "transformers>=4.40,<6",
]

ray = [
    "ray[default,train,tune]>=2.9",
]

dev = [
    "pytest>=8",
    "ruff>=0.8",
    "mypy>=1.0",
    "pre-commit>=4.0",
]

notebooks = [
    "jupyter>=1.0",
    "ipykernel>=6.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 200
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # import sorting
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"